requirements.txt
arabic-reshaper==2.1.4
asgiref==3.6.0
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
cssselect2==0.7.0
Django==4.1.5
djangorestframework==3.14.0
future==0.18.2
html5lib==1.1
idna==3.4
lxml==4.9.2
oscrypto==1.3.0
Pillow==9.4.0
pycparser==2.21
pyHanko==0.16.0
pyhanko-certvalidator==0.19.8
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.12
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
webencodings==0.5.1
xhtml2pdf==0.2.8
